+++ rfsocinterface/core/data/__init__.py
@@ -1,61 +1,101 @@
 """Submodule for data storage / processing related code."""
 
-from rfsocinterface.core.data.data import (
-    ProcessedData,
-    DEFAULT_MAP_DPIX,
-    N_POLARIZATION,
-    DECIMATE_ORDER,
-    BUTTER_ORDER,
-    PsdBasis,
-    flag_outliers,
-    plot_map,
-)
+
+from rfsocinterface.core.data.pipeline import Pipeline
 from rfsocinterface.core.data.routines import (
     ROUTINE_REGISTRY,
-    register_routine,
-    DataRoutine,
-    ProcessingStage,
+    BinTODIntoMap,
     CleanTOD,
+    ComputeNoisePSD,
+    CutoffFilter,
+    DataRoutine,
     HighPassFilter,
     LowPassFilter,
+    PlotMap,
+    PlotPSD,
+    ProcessingStage,
     RemoveElectronicsNoise,
-    BinTODIntoMap,
-    PlotMap
+    compute_templates,
+    decode_tone_indices,
+    find_peaks,
+    plot_psd_dbc_hz,
+    plot_psd_df_over_f,
+    register_routine,
 )
-# from rfsocinterface.core.data.map import (
-#     BinTODIntoMap,
-# )
-from rfsocinterface.core.data.pipeline import Pipeline
-from rfsocinterface.core.params import PARAM_FILE_N_TONE_ATTRIBUTES, initialize_params_file, update_params_file
+from rfsocinterface.core.data.storage import (
+    NewDataStorage,
+    ConsolidatedData,
+    ProcessedData,
+)
+from rfsocinterface.core.data.utils import (
+    AZ_TRIM,
+    ZA_TRIM,
+    CALIBRATION_TABLE_DTYPE,
+    DEFAULT_MAP_DPIX,
+    DECIMATE_ORDER,
+    N_POLARIZATION,
+    OPTCAM_OFFSET_AZ_PIX,
+    OPTCAM_OFFSET_ZA_PIX,
+    OPTCAM_PIX_SIZE_DEGREES,
+    RFSOC_TIME_OFFSET,
+    TONES_TABLE_DTYPE,
+    PsdBasis,
+    flag_outliers,
+    get_channel_group_name,
+    get_step_group_name,
+    generate_calibrated_data,
+    plot_map,
+    rotate_basis,
+)
 
-ROUTINE_NAME_MAP = {
-    'CleanTOD': CleanTOD,
-    'HighPassFilter': HighPassFilter,
-    'LowPassFilter': LowPassFilter,
-    'RemoveElectronicsNoise': RemoveElectronicsNoise,
-    # 'BinTODIntoMap': BinTODIntoMap,
-}
+# ROUTINE_NAME_MAP = {
+#     'CleanTOD': CleanTOD,
+#     'HighPassFilter': HighPassFilter,
+#     'LowPassFilter': LowPassFilter,
+#     'RemoveElectronicsNoise': RemoveElectronicsNoise,
+#     # 'BinTODIntoMap': BinTODIntoMap,
+# }
+ROUTINE_NAME_MAP = ROUTINE_REGISTRY  # Temporary fix until further edits made
 
 __all__ = [
-    ProcessedData,
+    AZ_TRIM,
+    ZA_TRIM,
+    CALIBRATION_TABLE_DTYPE,
+    DEFAULT_MAP_DPIX,
+    DECIMATE_ORDER,
+    N_POLARIZATION,
+    OPTCAM_OFFSET_AZ_PIX,
+    OPTCAM_OFFSET_ZA_PIX,
+    OPTCAM_PIX_SIZE_DEGREES,
+    ROUTINE_REGISTRY,
+    RFSOC_TIME_OFFSET,
+    TONES_TABLE_DTYPE,
+    BinTODIntoMap,
     CleanTOD,
+    ComputeNoisePSD,
+    ConsolidatedData,
+    CutoffFilter,
+    DataRoutine,
     HighPassFilter,
     LowPassFilter,
-    RemoveElectronicsNoise,
-    BinTODIntoMap,
+    NewDataStorage,
     Pipeline,
+    PlotMap,
+    PlotPSD,
+    ProcessedData,
     ProcessingStage,
-    DataRoutine,
     PsdBasis,
-    ROUTINE_REGISTRY,
-    DEFAULT_MAP_DPIX,
-    N_POLARIZATION,
-    PARAM_FILE_N_TONE_ATTRIBUTES,
-    DECIMATE_ORDER,
-    BUTTER_ORDER,
-    register_routine,
-    initialize_params_file,
-    update_params_file,
+    RemoveElectronicsNoise,
+    compute_templates,
+    decode_tone_indices,
+    find_peaks,
     flag_outliers,
+    get_channel_group_name,
+    get_step_group_name,
+    generate_calibrated_data,
     plot_map,
+    plot_psd_dbc_hz,
+    plot_psd_df_over_f,
+    register_routine,
+    rotate_basis,
 ]
