+++ rfsocinterface/core/data/routines.py
@@ -18,14 +18,16 @@ import datetime
 import json
 import h5py
 import matplotlib as mpl
+
+from rfsocinterface.core.data.storage import ConsolidatedData, ProcessedData
 mpl.use('QtAgg')
 import matplotlib.pyplot as plt
 from matplotlib.backends.backend_pdf import PdfPages
 from matplotlib.figure import Figure
 
 
-from rfsocinterface.core.data.data import ConsolidatedData, ProcessedData, PsdBasis, generate_calibrated_data, get_channel_group_name, get_step_group_name, rotate_basis, DEFAULT_MAP_DPIX, N_POLARIZATION, OPTCAM_PIX_SIZE_DEGREES, OPTCAM_OFFSET_AZ_PIX, OPTCAM_OFFSET_ZA_PIX
-from rfsocinterface.core.data.data import DECIMATE_ORDER
+from rfsocinterface.core.data.utils import PsdBasis, generate_calibrated_data, get_channel_group_name, get_step_group_name, rotate_basis, DEFAULT_MAP_DPIX, N_POLARIZATION, OPTCAM_PIX_SIZE_DEGREES, OPTCAM_OFFSET_AZ_PIX, OPTCAM_OFFSET_ZA_PIX
+from rfsocinterface.core.data.utils import DECIMATE_ORDER
 from rfsocinterface.core.utils import BUTTER_ORDER, GAUSSIAN_SIGMA, gaussian_filter, axis_index, get_git_hash, PERMISSIONS_ALL_FULL, axis_slice
 
 __all__ = (
@@ -1242,6 +1244,41 @@ class PlotMap(DataRoutine):
             fig.savefig(path, bbox_inches='tight')
         if self.params['show']:
             plt.show()
+
+
+def find_peaks(data: ProcessedData, primary_direction: str='az'):
+    import numpy as np
+    from numpy.polynomial import Polynomial
+    # find peak going forward / back
+    # fit gaussian
+    # take position of both peask
+    # right is 10-15
+    # left is 20-25
+    i_res = 241
+    right_indices = np.argwhere(np.logical_and(10 <= data.time, data.time <= 15)).flatten()
+    left_indices = np.argwhere(np.logical_and(20 <= data.time, data.time <= 25)).flatten()
+    telescope_pos = data.detector_az[i_res] if primary_direction.lower() == 'az' else data.detector_za[i_res]
+
+    right_peak_idx = right_indices[np.argmax(data.data_mK[i_res, right_indices])]
+    left_peak_idx = left_indices[np.argmax(data.data_mK[i_res, left_indices])]
+
+    right_slice = slice(right_peak_idx - 2, right_peak_idx + 3)
+    left_slice = slice(left_peak_idx - 2, left_peak_idx + 3)
+
+    right_fit = Polynomial.fit(telescope_pos[right_slice], data.data_mK[i_res, right_slice], 2).convert()
+    left_fit = Polynomial.fit(telescope_pos[left_slice], data.data_mK[i_res, left_slice], 2).convert()
+
+    right_az_0 = (-1 * right_fit.coef[1]) / (2 * right_fit.coef[2])
+    left_az_0 = (-1 * left_fit.coef[1]) / (2 * left_fit.coef[2])
+    plt.plot(telescope_pos[:], data.data_mK[i_res, :], label=f'Full Trace')
+    plt.plot(telescope_pos[right_slice], data.data_mK[i_res, right_slice], label=f'Right {primary_direction.upper()}_0 = {right_az_0}')
+    plt.plot(telescope_pos[left_slice], data.data_mK[i_res, left_slice], label=f'Left {primary_direction.upper()}_0 = {left_az_0}')
+    scan_rate = (telescope_pos[right_peak_idx + 10] - telescope_pos[right_peak_idx - 10]) \
+        / (data.time[right_peak_idx + 10] - data.time[right_peak_idx - 10])
+    time_delay = (left_az_0 - right_az_0) / scan_rate / 2  # Amount RFSoC is behind the telescope
+    plt.annotate(f'Time Delay (seconds RFSoC lags behind telescope)= {time_delay:.3f}s', (.1, .1), xycoords='axes fraction')
+    plt.legend()
+    plt.show()
 
 
 if __name__ == '__main__':
